requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
certifi==2022.6.15
charset-normalizer==2.1.0
Django==4.0.6
django-ckeditor==6.4.2
django-js-asset==2.0.0
django-modeltranslation==0.18.2
idna==3.3
mysqlclient==2.1.1
Pillow==9.2.0
requests==2.28.1
six==1.16.0
sqlparse==0.4.2
urllib3==1.26.10
